Update vector library (#33)
* Rename vec_2/ to vec2/

* Use float instead of double

* Improve Vector2 class

* Replace old Vec2 class

* Remove brace initialization

* Rename Vector2 to Vec2

* Rename Vec2 to Vector2

+++ src/entities/entities.h
@@ -20,7 +20,7 @@ class Entities {
 
   void input(SDL_Event event);
   void update();
-  void render(SDL_Renderer* renderer, double delay);
+  void render(SDL_Renderer* renderer, float delay);
 
  private:
   // static const int kMaxSize = 10000;
